+++ app.js
@@ -1,71 +1,71 @@
-"use strict";
+// "use strict";
 
-const express = require("express");
-const bodyParser = require("body-parser");
-const bcrypt = require("bcrypt-nodejs");
-const cors = require("cors");
-const knex = require("knex");
-const morgan = require("morgan");
-const path = require("path");
+// const express = require("express");
+// const bodyParser = require("body-parser");
+// const bcrypt = require("bcrypt-nodejs");
+// const cors = require("cors");
+// const knex = require("knex");
+// const morgan = require("morgan");
+// const path = require("path");
 
-const register = require("./controllers/register");
-const signin = require("./controllers/signin");
-const profile = require("./controllers/profile");
-const image = require("./controllers/image");
-const auth = require("./controllers/auth");
+// const register = require("./controllers/register");
+// const signin = require("./controllers/signin");
+// const profile = require("./controllers/profile");
+// const image = require("./controllers/image");
+// const auth = require("./controllers/auth");
 
-const { POSTGRES_URI } = require("./secret");
+// const { POSTGRES_URI } = require("./secret");
 
-const whitelist = [
-  "https://www.alonfabio.com",
-  "https://multitasker.alonfabio.com",
-  "http://multitasker.alonfabio.com",
-];
-const corsOptions = {
-  origin: function (origin, callback) {
-    if (whitelist.indexOf(origin) !== -1 || !origin) {
-      callback(null, true);
-    } else {
-      callback(new Error("Not allowed by CORS " + origin));
-    }
-  },
-};
+// const whitelist = [
+//   "https://www.alonfabio.com",
+//   "https://multitasker.alonfabio.com",
+//   "http://multitasker.alonfabio.com",
+// ];
+// const corsOptions = {
+//   origin: function (origin, callback) {
+//     if (whitelist.indexOf(origin) !== -1 || !origin) {
+//       callback(null, true);
+//     } else {
+//       callback(new Error("Not allowed by CORS " + origin));
+//     }
+//   },
+// };
 
-const db = knex({
-  // connect to your own database here.
-  client: "postgres",
-  connection: process.env.POSTGRES_URI || POSTGRES_URI,
-});
+// const db = knex({
+//   // connect to your own database here.
+//   client: "postgres",
+//   connection: process.env.POSTGRES_URI || POSTGRES_URI,
+// });
 
-const app = express();
+// const app = express();
 
-app.use(morgan("combined"));
-app.use(cors(corsOptions));
-app.use(bodyParser.json());
-app.use(express.static(path.join(__dirname, "build")));
+// app.use(morgan("combined"));
+// app.use(cors(corsOptions));
+// app.use(bodyParser.json());
+// app.use(express.static(path.join(__dirname, "build")));
 
-app.get("/", (req, res) => {
-  res.sendFile(path.join(__dirname, "build", "index.html"));
-});
-app.post("/signin", signin.signinAuthentication(db, bcrypt));
-app.post("/register", (req, res) => {
-  register.handleRegister(req, res, db, bcrypt);
-});
-app.get("/profile/:id", auth.getAuthentication, (req, res) => {
-  profile.handleProfileGet(req, res, db);
-});
-app.post("/profile/:id", auth.getAuthentication, (req, res) => {
-  profile.handleProfilePost(req, res, db);
-});
-app.put("/image", auth.getAuthentication, (req, res) => {
-  image.handleImage(req, res, db);
-});
-app.post("/imageurl", auth.getAuthentication, (req, res) => {
-  image.handleApiCall(req, res);
-});
+// app.get("/", (req, res) => {
+//   res.sendFile(path.join(__dirname, "build", "index.html"));
+// });
+// app.post("/signin", signin.signinAuthentication(db, bcrypt));
+// app.post("/register", (req, res) => {
+//   register.handleRegister(req, res, db, bcrypt);
+// });
+// app.get("/profile/:id", auth.getAuthentication, (req, res) => {
+//   profile.handleProfileGet(req, res, db);
+// });
+// app.post("/profile/:id", auth.getAuthentication, (req, res) => {
+//   profile.handleProfilePost(req, res, db);
+// });
+// app.put("/image", auth.getAuthentication, (req, res) => {
+//   image.handleImage(req, res, db);
+// });
+// app.post("/imageurl", auth.getAuthentication, (req, res) => {
+//   image.handleApiCall(req, res);
+// });
 
 // app.listen(8080, () => {
 //   console.log("app is running on port 8080");
 // });
 
-module.exports = app;
+// // module.exports = app;
